requirements.txt
fastapi
uvicorn
motor
pymongo
passlib
python-dotenv
pyjwt
